requirements.txt
bcrypt==4.2.1
cryptography==44.0.0
pyotp==2.9.0
psycopg2-binary==2.9.10
qrcode==8.0
python-dotenv==1.0.1
pillow==11.0.0
